requirements.txt
streamlit
pytesseract
Pillow
openai
python-dotenv
